+++ src/ipc_listener.hpp
@@ -56,9 +56,6 @@ class ipc_listener_t : public stream_listener_base_t
     //  Handlers for I/O events.
     void in_event ();
 
-    // Create wildcard path address
-    static int create_wildcard_address (std::string &path_, std::string &file_);
-
     //  Filter new connections if the OS provides a mechanism to get
     //  the credentials of the peer process.  Called from accept().
 #if defined ZMQ_HAVE_SO_PEERCRED || defined ZMQ_HAVE_LOCAL_PEERCRED
@@ -81,9 +78,6 @@ class ipc_listener_t : public stream_listener_base_t
 
     //  Name of the file associated with the UNIX domain address.
     std::string _filename;
-
-    // Acceptable temporary directory environment variables
-    static const char *tmp_env_vars[];
 
     ipc_listener_t (const ipc_listener_t &);
     const ipc_listener_t &operator= (const ipc_listener_t &);
